# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TechSwiftTrix ERP API
  version: 1.0.0
  description: |
    Comprehensive REST API for the TechSwiftTrix ERP System — a multi-tenant enterprise
    resource platform supporting operations across Africa. Covers authentication, user
    management, client management, payments, contracts, and audit logging.
  contact:
    email: support@techswifttrix.com
  license:
    name: Proprietary
    url: https://techswifttrix.com
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://staging-api.techswifttrix.com
    description: Staging
  - url: https://api.techswifttrix.com
    description: Production
security:
  - BearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@techswifttrix.com
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout and invalidate session
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /clients:
    get:
      tags: [Clients]
      summary: List clients (paginated, filterable)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [lead, active, inactive, churned]
          description: Filter by client status
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
      responses:
        '200':
          description: Paginated list of clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
            example:
              firstName: John
              lastName: Kamau
              email: john.kamau@example.com
              phone: "+254712345678"
              company: Kamau Enterprises
              country: KE
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Validation failed
            details: [{ field: email, message: "must be a valid email" }]
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT access token
        expiresIn:
          type: integer
          description: Token expiry in seconds
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [CEO, CoS, CFO, COO, CTO, EA, HeadOfTrainers, Trainer, Agent, OperationsUser, TechStaff, Developer]
...
    Client:
      type: object
      required: [id, firstName, lastName, email, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code
        status:
          type: string
          enum: [lead, active, inactive, churned]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClientResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Client'
    ClientListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Client'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateClientRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        country:
          type: string
...